# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/comment/upvotes": {
            "post": {
                "tags": [
                    "CommentVote"
                ],
                "summary": "user likes a comment",
                "description": "A user likes a comment",
                "operationId": "b03341e1472eeb109dfcfc533f24783a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "post_id",
                                    "like",
                                    "dislike"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "required|integer"
                                    },
                                    "post_id": {
                                        "type": "required|integer"
                                    },
                                    "like": {
                                        "type": "required|boolean"
                                    },
                                    "dislike": {
                                        "type": "required|boolean|default(0)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "You like this comment",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/post/downvotes": {
            "post": {
                "tags": [
                    "PostVote"
                ],
                "summary": "user dislikes a comment",
                "description": "A user dislikes a comment",
                "operationId": "7bedc600e37d51a87bf67c57c4853abf",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "post_id",
                                    "like",
                                    "dislike"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "required|integer"
                                    },
                                    "post_id": {
                                        "type": "required|integer"
                                    },
                                    "like": {
                                        "type": "required|boolean"
                                    },
                                    "dislike": {
                                        "type": "required|boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "You dislike this comment",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/comment/votes/{post_id}": {
            "get": {
                "tags": [
                    "CommentVote"
                ],
                "summary": "Get count of votes of comments",
                "description": "Get count of votes of comments",
                "operationId": "0cb3a031bf298734ccc9360e3f26ed3a",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to get count of votes"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/upvotes": {
            "post": {
                "tags": [
                    "PostVote"
                ],
                "summary": "user likes a post",
                "description": "A user likes a post",
                "operationId": "7a8de6f95d67444acb7f66df941b7030",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "post_id",
                                    "like",
                                    "dislike"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "required|integer"
                                    },
                                    "post_id": {
                                        "type": "required|integer"
                                    },
                                    "like": {
                                        "type": "required|boolean"
                                    },
                                    "dislike": {
                                        "type": "required|boolean|default(0)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "You like this post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/post/votes/{post_id}": {
            "get": {
                "tags": [
                    "PostVote"
                ],
                "summary": "Get count of votes from post",
                "description": "Get count of votes from post",
                "operationId": "43506f45b4af14e6a60f42a501299a50",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to get count of votes"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}